New 'attributes' phone number feature
Allows specification of group, class, and traffic type for phone
numbers. The 'group' and 'class' parameters are arbitrary labels
that can be used in future for internal or external apps;
the 'traffic' parameter will initially be used for setting T.38
fax variables on fax calls when it is set to value of 'fax'.