Added toggle to agent queue login callflow.
A 'toggle' action allows a single BLF button to be used for logging
an agent in/out of a queue. If the agent is in any logged-in state
then they will be logged out and vice versa. The exception to that
is if the agent is logged in and paused, they will be unpaused
instead of logged out.

There is also a 'toggle_paused' action which toggles paused/unpaused
exclusively.